+++ src/auth/apiKey.js
@@ -21,9 +21,13 @@ export default class ApiKey extends Auth {
     };
   }
 
-  async logout(credentials) {
-    await super.logout(credentials);
-    if (this.router.currentRoute.value.name !== 'logout') {
+  async logout(credentials, quiet = false) {
+    await super.logout(credentials, quiet);
+    if (quiet) {
+      // Other schemes are still logged in, stay on the current page
+      await this.confirmLogout();
+    }
+    else if (this.router.currentRoute.value.name !== 'logout') {
       this.router.push('/auth/logout');
     }
     return true;